onWatced onQueue - разбил на мелкие функции
По сути их потом объединить в одну ф-цию.
Не критичное обновление, для изящности кода.
