+++ src/js/listeners/onWatched.js
@@ -2,37 +2,24 @@ import { refs } from '../base/refs';
 import renderCardsHero from '../base/renderCardsHero';
 import renderPagination from '../base/renderPagination';
 import onLibraryChangeClassBtnWatch from '../changeClassHeder/changeClassLibraryBtnWatch';
+import isArrEmpty from '../base/isArrEmpty';
+import putSettings from '../base/putSettings';
+import paginationDataCalc from '../base/paginationDataCalc';
 
 // Кнопка Watched в хедері
 export default function onWatched(e, page = 1) {
   onLibraryChangeClassBtnWatch();
   refs.hero.innerHTML = '';
   refs.empty.classList.add('hidden');
-
   const PAGE_SIZE = 20;
   let arr = JSON.parse(localStorage.getItem('library-watched'));
 
-  if (!arr || !arr.length) {
-    refs.pagination.classList.add('hidden');
-    refs.hero.classList.add('library-foto');
-    refs.empty.classList.remove('hidden');
-    console.log('not data');
-    return;
-  }
+  if (isArrEmpty(arr, refs)) return; // если нет карточек
 
-  const settings = {
-    page: page,
-    fetch: 'Watched',
-  };
+  putSettings(page, "Watched");
 
-  localStorage.setItem('page', JSON.stringify(settings));
+  const [newArr, totalPage] = paginationDataCalc(arr, page, PAGE_SIZE);
 
-  const totalPage = Math.ceil(arr.length / PAGE_SIZE);
-  const position = (page - 1) * PAGE_SIZE;
-
-  arr = arr.splice(position, PAGE_SIZE);
-  localStorage.setItem('cards', JSON.stringify(arr));
-
-  renderCardsHero(arr);
+  renderCardsHero(newArr);
   renderPagination(totalPage, page);
 }
